move signedState to state package
and make public (inc. methods)
