+++ channel/state/signedstate.go
@@ -1,26 +1,24 @@
-package channel
+package state
 
 import (
 	"errors"
 	"fmt"
-
-	"github.com/statechannels/go-nitro/channel/state"
 )
 
-type signedState struct {
-	state state.State
-	sigs  map[uint]state.Signature // keyed by participant index
+type SignedState struct {
+	state State
+	sigs  map[uint]Signature // keyed by participant index
 }
 
-// newSignedState initializes a SignedState struct for the given state.
+// newSignedState initializes a SignedState struct for the given
 // The signedState returned will have no signatures.
-func newSignedState(s state.State) signedState {
-	return signedState{s, make(map[uint]state.Signature, len(s.Participants))}
+func NewSignedState(s State) SignedState {
+	return SignedState{s, make(map[uint]Signature, len(s.Participants))}
 }
 
-// addSignature adds a participant's signature for the state.
+// addSignature adds a participant's signature for the
 // An error is thrown if the signature is invalid.
-func (ss signedState) addSignature(sig state.Signature) error {
+func (ss SignedState) AddSignature(sig Signature) error {
 	signer, err := ss.state.RecoverSigner(sig)
 	if err != nil {
 		return fmt.Errorf("addSignature failed to recover signer %w", err)
@@ -42,18 +40,18 @@ func (ss signedState) addSignature(sig state.Signature) error {
 	return errors.New("signature does not match any participant")
 
 }
-func (ss signedState) State() state.State {
+func (ss SignedState) State() State {
 	return ss.state
 }
 
 // HasSignatureForParticipant returns true if the participant (at participantIndex) has a valid signature.
-func (ss signedState) HasSignatureForParticipant(participantIndex uint) bool {
+func (ss SignedState) HasSignatureForParticipant(participantIndex uint) bool {
 	_, found := ss.sigs[uint(participantIndex)]
 	return found
 }
 
 // HasAllSignatures returns true if every participant has a valid signature.
-func (ss signedState) HasAllSignatures() bool {
+func (ss SignedState) HasAllSignatures() bool {
 	// Since signatures are validated
 	if len(ss.sigs) == len(ss.state.Participants) {
 		return true
